change float to double

+++ src/test/FILEINFO_FITS_MULTIHDU.test.ts
@@ -47,7 +47,7 @@ let assertItem: AssertItem = {
                 stokesVals: [],
                 computedEntries: [
                     { name: "Name", value: "spire500_ext.fits" },
-		    { name: "Data type", value: "float" },		    
+		    { name: "Data type", value: "double" },
                     { name: 'HDU', value: '1' },
                     { name: 'Extension name', value: 'image' },
                     { name: "Shape", value: "[830, 870]" },
@@ -147,7 +147,7 @@ let assertItem: AssertItem = {
                 stokesVals: [],
                 computedEntries: [
                     { name: "Name", value: "spire500_ext.fits" },
-		    { name: "Data type", value: "float" },		    
+		    { name: "Data type", value: "double" },
                     { name: 'HDU', value: '6' },
                     { name: 'Extension name', value: 'error' },
                     { name: "Shape", value: "[830, 870]" },
@@ -247,7 +247,7 @@ let assertItem: AssertItem = {
                 stokesVals: [],
                 computedEntries: [
                     { name: "Name", value: "spire500_ext.fits" },
-		    { name: "Data type", value: "float" },		    
+		    { name: "Data type", value: "double" },
                     { name: 'HDU', value: '7' },
                     { name: 'Extension name', value: 'coverage' },
                     { name: "Shape", value: "[830, 870]" },
